backend/requirements.txt
Flask
torch==1.12.1+cpu
torchvision==0.13.1+cpu
Pillow
flask-cors
